Imdb_Review_Rating_ML_MODEL
The dataset consists of 5000+ IMDB movies and related information scraped from various scources and all the data are combined to form a Dataset which can be used to train a machine learning model to predict the ratings based on pre set criteria.It has 27 independent variables and one dependent(target) variable. The dataset has information about 4998 unique movies, its directors and actors. In total there are 2399 directors, 8161 actors and their popularity. The combined dataset is available on Github.